run_qc.py: move code into the 'cli.run_qc' submodule.
